# Speed up point_to_coords