- if cache doesn't exist, find coord when making cache
- else, use binary search