@kngwyu 2018-07-18
Now it looks much smarter, thanks @birkenfeld 

Here's the result of std benchmark:
before
```
running 3 tests
test completes_hashmap               ... bench:     798,103 ns/iter (+/- 43,783)
test completes_methods_for_file_open ... bench:     772,886 ns/iter (+/- 32,676)
test completes_methods_for_vec       ... bench: 112,440,661 ns/iter (+/- 5,331,488)
```
after
```
running 3 tests
test completes_hashmap               ... bench:     808,920 ns/iter (+/- 25,661)
test completes_methods_for_file_open ... bench:     786,085 ns/iter (+/- 45,139)
test completes_methods_for_vec       ... bench: 109,354,725 ns/iter (+/- 1,658,401)
```