@birkenfeld 2018-07-18
Doesn't look much faster...?